Refactor icon utils and reuse in source panel
